// src/crmUi/prospectos/pages/sharedForm/types/prospectosFormValidation.ts
export interface ProspectosFormValidation {
	isRazonSocialRequired: boolean
	isContactoRequired: boolean
	isCorreoRequired: boolean
	isValidCorreo: boolean
	isTelefonoRequired: boolean
	isUenRequired: boolean
	isSegmentoRequired: boolean
	isTipoClienteRequired: boolean
	isVPORequired: boolean
	isTerritorioRequired: boolean
	isObservacionesRequired: boolean
}